feat(): fixed cors issues

+++ model/model.go
@@ -8,6 +8,6 @@ type Post struct {
 	Name      string    `gorm:"type:varchar(128);not null" json:"name"`
 	Company   string    `gorm:"type:varchar(32)" json:"company"`
 	Section   string    `gorm:"type:varchar(32)" json:"section"`
-	ItemType  int       `gorm:"type:bit(1)" json:"itemType"`
+	ItemType  uint8     `gorm:"type:bit(1)" json:"itemType"`
 	CreatedAt time.Time `gorm:"default:CURRENT_TIMESTAMP" json:"created_at"`
 }
